Employee deleted successfully!

+++ app/Http/Services/EmployeeService.php
@@ -49,6 +49,10 @@ class EmployeeService
     {
         DB::beginTransaction();
         try {
+            $employee = $this->employeeRepositoryInterface->find($id);
+            if (!$employee) {
+                throw new HttpException(404, 'Employee not found!');
+            }
             $data = $this->employeeRepositoryInterface->delete($id);
             DB::commit();
             return $data;
